fix: add refresh on ticket Stopwatch
