+++ packages/apps/helpdesk/src/components/templates/index.js
@@ -24,5 +24,6 @@ export {default as TicketEditButton} from './TicketEditButton/TicketEditButton';
 export {default as TicketHeader} from './TicketHeader/TicketHeader';
 export {default as TicketSearchBar} from './TicketSearchBar/TicketSearchBar';
 export {default as TicketsStatusButton} from './TicketsStatusButton/TicketsStatusButton';
+export {default as TicketStopwatch} from './TicketStopwatch/TicketStopwatch';
 export {default as TicketTypeSearchBar} from './TicketTypeSearchBar/TicketTypeSearchBar';
 export {default as UserSearchBar} from './UserSearchBar/UserSearchBar';
